Blogpost_generator/requirements.txt
langgraph>=0.0.11
langchain>=0.1.0
langsmith>=0.0.57
python-dotenv>=1.0.0
transformers>=4.36.0
torch>=2.1.0
groq>=0.4.2
google-search-results>=2.4.2
streamlit>=1.32.0 